Fix bug with toggleRead on cached books
This commit fixes a bug with books that were loaded in from local storage. Previously, if local storage was loaded in,
you would not be able to use the toggleRead function from Book.prototype.

Thankfully, when we load in the cached books, we can just add that property back to fix the issue.